Create index.test.js for sample app too
